Fix image_index type to size_t and add image_list::npos

+++ src/xtd.forms/src/xtd/forms/tab_page.cpp
@@ -27,11 +27,11 @@ scrollable_control& tab_page::auto_scroll(bool value) {
   return *this;
 }
 
-int32_t tab_page::image_index() const noexcept {
+size_t tab_page::image_index() const noexcept {
   return data_->image_index;
 }
 
-tab_page& tab_page::image_index(int32_t value) {
+tab_page& tab_page::image_index(size_t value) {
   if (data_->image_index != value) {
     data_->image_index = value;
     if (is_handle_created() && parent().has_value()) native::tab_page::image_index(handle(), data_->image_index);
